+++ transformer_engine/pytorch/transformer.py
@@ -14,6 +14,8 @@ from transformer_engine.pytorch.module import LayerNormMLP, LayerNorm, RMSNorm
 from transformer_engine.pytorch.attention import (
     InferenceParams,
     MultiheadAttention,
+)
+from transformer_engine.pytorch.dot_product_attention.utils import (
     check_set_window_size,
 )
 from transformer_engine.pytorch.jit import (
